Tidy up and treat CA1034 as an error.  Fix remaining nested classes.
