+++ Terminal.Gui/Drawing/LineCanvas.cs
@@ -213,9 +213,9 @@ namespace Terminal.Gui {
 		/// so that all lines connect up with the appropriate intersection symbols. 
 		/// </summary>
 		/// <returns>A map of all the points within the canvas.</returns>
-		public Dictionary<Point, Cell> GetCellMap ()
+		public Dictionary<Point, LineCanvasCell> GetCellMap ()
 		{
-			var map = new Dictionary<Point, Cell> ();
+			var map = new Dictionary<Point, LineCanvasCell> ();
 
 			// walk through each pixel of the bitmap
 			for (int y = Bounds.Y; y < Bounds.Y + Bounds.Height; y++) {
@@ -528,29 +528,13 @@ namespace Terminal.Gui {
 
 		}
 
-		/// <summary>
-		/// Represents a single row/column within the <see cref="LineCanvas"/>. Includes the glyph and the foreground/background colors.
-		/// </summary>
-		public class Cell {
-			/// <summary>
-			/// The glyph to draw.
-			/// </summary>
-			public Rune? Rune { get; set; }
-
-			/// <summary>
-			/// The foreground color to draw the glyph with.
-			/// </summary>
-			public Attribute? Attribute { get; set; }
-
-		}
-
-		private Cell GetCellForIntersects (ConsoleDriver driver, IntersectionDefinition [] intersects)
+		private LineCanvasCell GetCellForIntersects (ConsoleDriver driver, IntersectionDefinition [] intersects)
 		{
 			if (!intersects.Any ()) {
 				return null;
 			}
 
-			var cell = new Cell ();
+			var cell = new LineCanvasCell ();
 			cell.Rune = GetRuneForIntersects (driver, intersects);
 			cell.Attribute = GetAttributeForIntersects (intersects);
 			return cell;
