+++ Terminal.Gui/Drawing/LineCanvasCell.cs
@@ -1,0 +1,21 @@
+﻿using System.Text;
+
+
+namespace Terminal.Gui {
+
+	/// <summary>
+	/// Represents a single row/column within the <see cref="LineCanvas"/>. Includes the glyph and the foreground/background colors.
+	/// </summary>
+	public class LineCanvasCell {
+		/// <summary>
+		/// The glyph to draw.
+		/// </summary>
+		public Rune? Rune { get; set; }
+
+		/// <summary>
+		/// The foreground color to draw the glyph with.
+		/// </summary>
+		public Attribute? Attribute { get; set; }
+
+	}
+}
