+++ Terminal.Gui/Views/TabMouseEventArgs.cs
@@ -13,7 +13,7 @@ namespace Terminal.Gui {
 		/// </summary>
 		/// <remarks>This will be null if the click is after last tab
 		/// or before first.</remarks>
-		public TabView.Tab Tab { get; }
+		public Tab Tab { get; }
 
 		/// <summary>
 		/// Gets the actual mouse event.  Use <see cref="MouseEvent.Handled"/> to cancel this event
@@ -26,7 +26,7 @@ namespace Terminal.Gui {
 		/// </summary>
 		/// <param name="tab"><see cref="TabView.Tab"/> that the mouse was over when the event occurred.</param>
 		/// <param name="mouseEvent">The mouse activity being reported</param>
-		public TabMouseEventArgs (TabView.Tab tab, MouseEvent mouseEvent)
+		public TabMouseEventArgs (Tab tab, MouseEvent mouseEvent)
 		{
 			Tab = tab;
 			MouseEvent = mouseEvent;
