+++ UICatalog/Scenarios/Editor.cs
@@ -751,9 +751,9 @@ namespace UICatalog.Scenarios {
 				Height = Dim.Fill ()
 			};
 
-			_tabView.AddTab (new TabView.Tab ("Find", FindTab ()), isFind);
+			_tabView.AddTab (new Tab ("Find", FindTab ()), isFind);
 			var replace = ReplaceTab ();
-			_tabView.AddTab (new TabView.Tab ("Replace", replace), !isFind);
+			_tabView.AddTab (new Tab ("Replace", replace), !isFind);
 			_tabView.SelectedTabChanged += (s, e) => _tabView.SelectedTab.View.FocusFirst ();
 			_winDialog.Add (_tabView);
