+++ UICatalog/Scenarios/TabViewExample.cs
@@ -59,18 +59,18 @@ namespace UICatalog.Scenarios {
 				Height = 20,
 			};
 
-			tabView.AddTab (new TabView.Tab ("Tab1", new Label ("hodor!")), false);
-			tabView.AddTab (new TabView.Tab ("Tab2", new Label ("durdur")), false);
-			tabView.AddTab (new TabView.Tab ("Interactive Tab", GetInteractiveTab ()), false);
-			tabView.AddTab (new TabView.Tab ("Big Text", GetBigTextFileTab ()), false);
-			tabView.AddTab (new TabView.Tab (
+			tabView.AddTab (new Tab ("Tab1", new Label ("hodor!")), false);
+			tabView.AddTab (new Tab ("Tab2", new Label ("durdur")), false);
+			tabView.AddTab (new Tab ("Interactive Tab", GetInteractiveTab ()), false);
+			tabView.AddTab (new Tab ("Big Text", GetBigTextFileTab ()), false);
+			tabView.AddTab (new Tab (
 				"Long name Tab, I mean seriously long.  Like you would not believe how long this tab's name is its just too much really woooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooowwww thats long",
 				 new Label ("This tab has a very long name which should be truncated.  See TabView.MaxTabTextWidth")),
 				 false);
-			tabView.AddTab (new TabView.Tab ("Les Mise" + Char.ConvertFromUtf32 (Int32.Parse ("0301", NumberStyles.HexNumber)) + "rables", new Label ("This tab name is unicode")), false);
+			tabView.AddTab (new Tab ("Les Mise" + Char.ConvertFromUtf32 (Int32.Parse ("0301", NumberStyles.HexNumber)) + "rables", new Label ("This tab name is unicode")), false);
 
 			for (int i = 0; i < 100; i++) {
-				tabView.AddTab (new TabView.Tab ($"Tab{i}", new Label ($"Welcome to tab {i}")), false);
+				tabView.AddTab (new Tab ($"Tab{i}", new Label ($"Welcome to tab {i}")), false);
 			}
 
 			tabView.SelectedTab = tabView.Tabs.First ();
@@ -115,7 +115,7 @@ namespace UICatalog.Scenarios {
 
 		private void AddBlankTab ()
 		{
-			tabView.AddTab (new TabView.Tab (), false);
+			tabView.AddTab (new Tab (), false);
 		}
 
 		private View GetInteractiveTab ()
