+++ UnitTests/Views/TabViewTests.cs
@@ -23,15 +23,15 @@ namespace Terminal.Gui.ViewsTests {
 			return GetTabView (out _, out _);
 		}
 
-		private TabView GetTabView (out TabView.Tab tab1, out TabView.Tab tab2, bool initFakeDriver = true)
+		private TabView GetTabView (out Tab tab1, out Tab tab2, bool initFakeDriver = true)
 		{
 			if (initFakeDriver)
 				InitFakeDriver ();
 
 			var tv = new TabView ();
 			tv.ColorScheme = new ColorScheme ();
-			tv.AddTab (tab1 = new TabView.Tab ("Tab1", new TextField ("hi")), false);
-			tv.AddTab (tab2 = new TabView.Tab ("Tab2", new Label ("hi2")), false);
+			tv.AddTab (tab1 = new Tab ("Tab1", new TextField ("hi")), false);
+			tv.AddTab (tab2 = new Tab ("Tab2", new Label ("hi2")), false);
 			return tv;
 		}
 
@@ -41,10 +41,10 @@ namespace Terminal.Gui.ViewsTests {
 			InitFakeDriver ();
 
 			var tv = new TabView ();
-			TabView.Tab tab1;
-			TabView.Tab tab2;
-			tv.AddTab (tab1 = new TabView.Tab ("Tab1", new TextField ("hi")), false);
-			tv.AddTab (tab2 = new TabView.Tab ("Tab1", new Label ("hi2")), true);
+			Tab tab1;
+			Tab tab2;
+			tv.AddTab (tab1 = new Tab ("Tab1", new TextField ("hi")), false);
+			tv.AddTab (tab2 = new Tab ("Tab1", new Label ("hi2")), true);
 
 			Assert.Equal (2, tv.Tabs.Count);
 			Assert.Equal (tab2, tv.SelectedTab);
@@ -98,8 +98,8 @@ namespace Terminal.Gui.ViewsTests {
 
 			tv.SelectedTab = tab1;
 
-			TabView.Tab oldTab = null;
-			TabView.Tab newTab = null;
+			Tab oldTab = null;
+			Tab newTab = null;
 			int called = 0;
 
 			tv.SelectedTabChanged += (s, e) => {
@@ -172,13 +172,13 @@ namespace Terminal.Gui.ViewsTests {
 		{
 			var tv = GetTabView (out var tab1, out var tab2);
 
-			TabView.Tab tab3;
-			TabView.Tab tab4;
-			TabView.Tab tab5;
+			Tab tab3;
+			Tab tab4;
+			Tab tab5;
 
-			tv.AddTab (tab3 = new TabView.Tab (), false);
-			tv.AddTab (tab4 = new TabView.Tab (), false);
-			tv.AddTab (tab5 = new TabView.Tab (), false);
+			tv.AddTab (tab3 = new Tab (), false);
+			tv.AddTab (tab4 = new Tab (), false);
+			tv.AddTab (tab5 = new Tab (), false);
 
 			tv.SelectedTab = tab1;
 
@@ -320,7 +320,7 @@ namespace Terminal.Gui.ViewsTests {
 			var tv = GetTabView (out var tab1, out var tab2, false);
 			tv.Width = 10;
 			tv.Height = 5;
-			tv.Style = new TabView.TabStyle { ShowTopLine = false };
+			tv.Style = new TabStyle { ShowTopLine = false };
 			tv.ApplyStyleChanges ();
 
 			// Ensures that the tab bar subview gets the bounds of the parent TabView
@@ -414,7 +414,7 @@ namespace Terminal.Gui.ViewsTests {
 			var tv = GetTabView (out _, out _, false);
 			tv.Width = 4;
 			tv.Height = 5;
-			tv.Style = new TabView.TabStyle { ShowTopLine = false };
+			tv.Style = new TabStyle { ShowTopLine = false };
 			tv.ApplyStyleChanges ();
 			tv.LayoutSubviews ();
 
@@ -452,7 +452,7 @@ namespace Terminal.Gui.ViewsTests {
 			var tv = GetTabView (out _, out _, false);
 			tv.Width = 3;
 			tv.Height = 5;
-			tv.Style = new TabView.TabStyle { ShowTopLine = false };
+			tv.Style = new TabStyle { ShowTopLine = false };
 			tv.ApplyStyleChanges ();
 			tv.LayoutSubviews ();
 
@@ -472,7 +472,7 @@ namespace Terminal.Gui.ViewsTests {
 			var tv = GetTabView (out var tab1, out var tab2, false);
 			tv.Width = 10;
 			tv.Height = 5;
-			tv.Style = new TabView.TabStyle { TabsOnBottom = true };
+			tv.Style = new TabStyle { TabsOnBottom = true };
 			tv.ApplyStyleChanges ();
 
 			// Ensures that the tab bar subview gets the bounds of the parent TabView
@@ -535,7 +535,7 @@ namespace Terminal.Gui.ViewsTests {
 			var tv = GetTabView (out var tab1, out var tab2, false);
 			tv.Width = 10;
 			tv.Height = 5;
-			tv.Style = new TabView.TabStyle { ShowTopLine = false, TabsOnBottom = true };
+			tv.Style = new TabStyle { ShowTopLine = false, TabsOnBottom = true };
 			tv.ApplyStyleChanges ();
 
 			// Ensures that the tab bar subview gets the bounds of the parent TabView
@@ -611,7 +611,7 @@ namespace Terminal.Gui.ViewsTests {
 			var tv = GetTabView (out _, out _, false);
 			tv.Width = 4;
 			tv.Height = 5;
-			tv.Style = new TabView.TabStyle { TabsOnBottom = true };
+			tv.Style = new TabStyle { TabsOnBottom = true };
 			tv.ApplyStyleChanges ();
 			tv.LayoutSubviews ();
 
@@ -631,7 +631,7 @@ namespace Terminal.Gui.ViewsTests {
 			var tv = GetTabView (out _, out _, false);
 			tv.Width = 4;
 			tv.Height = 5;
-			tv.Style = new TabView.TabStyle { ShowTopLine = false, TabsOnBottom = true };
+			tv.Style = new TabStyle { ShowTopLine = false, TabsOnBottom = true };
 			tv.ApplyStyleChanges ();
 			tv.LayoutSubviews ();
 
@@ -651,7 +651,7 @@ namespace Terminal.Gui.ViewsTests {
 			var tv = GetTabView (out _, out _, false);
 			tv.Width = 3;
 			tv.Height = 5;
-			tv.Style = new TabView.TabStyle { TabsOnBottom = true };
+			tv.Style = new TabStyle { TabsOnBottom = true };
 			tv.ApplyStyleChanges ();
 			tv.LayoutSubviews ();
 
@@ -671,7 +671,7 @@ namespace Terminal.Gui.ViewsTests {
 			var tv = GetTabView (out _, out _, false);
 			tv.Width = 3;
 			tv.Height = 5;
-			tv.Style = new TabView.TabStyle { ShowTopLine = false, TabsOnBottom = true };
+			tv.Style = new TabStyle { ShowTopLine = false, TabsOnBottom = true };
 			tv.ApplyStyleChanges ();
 			tv.LayoutSubviews ();
 
@@ -724,7 +724,7 @@ namespace Terminal.Gui.ViewsTests {
 			var tv = GetTabView (out var tab1, out var tab2, false);
 			tv.Width = 20;
 			tv.Height = 5;
-			tv.Style = new TabView.TabStyle { TabsOnBottom = true };
+			tv.Style = new TabStyle { TabsOnBottom = true };
 			tv.ApplyStyleChanges ();
 
 			tv.LayoutSubviews ();
@@ -776,7 +776,7 @@ namespace Terminal.Gui.ViewsTests {
 └──────────────────┘
 ", output);
 
-			TabView.Tab clicked = null;
+			Tab clicked = null;
 			
 
 			tv.TabClicked += (s,e)=>{
